Great leaders are willing to sacrifice the numbers to save the people. Poor leaders sacrifice the people to save the numbers

+++ 0x05-pointers_arrays_strings/100-atoi.c
@@ -9,10 +9,10 @@
 
 int    _atoi(const char *s)
 {
-	int	i;
-	int	result;
-	int	sign;
-	int	j;
+	long int	i;
+	int		result;
+	int		sign;
+	int		j;
 
 	sign = 1;
 	i = 0;
